feat: implement GeneralNavBar and SpecificNavBar components for improved navigation
