+++ src/components/layout/navbar/specific/style.ts
@@ -1,23 +1,20 @@
 import styled from "styled-components";
 import { breakpoints } from "@styles/breakpoints";
 import { Link } from "react-router-dom";
-import type { LinkElementProps } from "./types";
+import type { LinkElementProps } from "../types";
 
 export const LayoutNavbar = styled.nav`
-  position: fixed;
-  top: 0;
+  position: relative;
   width: 100%;
   display: flex;
   align-items: center;
-  z-index: 1001;
-  padding: 1rem 1rem;
-  box-sizing: border-box;
   justify-content: space-between;
-  background-color: #00000050;
+  background-color: #8181811f;
 
   @media (min-width: ${breakpoints.lg}) {
     justify-content: flex-start;
-    background-color: transparent;
+      background-color: transparent;
+
   }
 `;
 
@@ -33,34 +30,10 @@ export const LogoLayout = styled.img`
 export const MenuLayout = styled.div`
   font-size: 2rem;
   cursor: pointer;
-  color: white;
+  color: black;
 
   @media (min-width: ${breakpoints.lg}) {
     display: none;
-  }
-`;
-
-export const AuxiliarMenu = styled.div<{ open: boolean }>`
-  display: ${({ open }) => (open ? "flex" : "none")};
-  flex-direction: column;
-  width: 100%;
-  opacity: ${({ open }) => (open ? "1" : "0")};
-
-  @media (min-width: ${breakpoints.lg}) {
-    display: flex;
-    flex-direction: row;
-    gap: 2rem;
-    border-radius: 2rem;
-    background-color: #111111c4;
-    padding: 0.75rem 0.4rem;
-
-    width: auto;
-    transform: none;
-    opacity: 1;
-    transition: none;
-    transform-origin: initial;
-    position: static;
-    visibility: visible;
   }
 `;
 
@@ -74,6 +47,8 @@ export const MenuElement = styled.ul<{ open: boolean }>`
   visibility: ${({ open }) => (open ? "visible" : "hidden")};
   transition: all 0.4s cubic-bezier(0.25, 0.8, 0.25, 1);
   transform-origin: top center;
+  margin: 0;
+  padding: 0;
 
   @media (min-width: ${breakpoints.lg}) {
     display: flex;
@@ -82,11 +57,8 @@ export const MenuElement = styled.ul<{ open: boolean }>`
     width: auto;
     left: auto;
     right: auto;
-    background: transparent;
     gap: 2rem;
     border-radius: 2rem;
-    background-color: #f5f5f5be;
-    padding: 0.75rem 0.4rem;
     transform: none;
     opacity: 1;
     visibility: visible;
@@ -105,9 +77,9 @@ export const LinkElement = styled(Link)<LinkElementProps>`
   font-optical-sizing: auto;
   font-weight: 500;
   font-style: normal;
-  color: white;
   padding: 1rem 2rem;
   text-align: center;
+  color: ${({ color }) => color || "black"};
 
   &::after {
     content: "";
@@ -116,10 +88,11 @@ export const LinkElement = styled(Link)<LinkElementProps>`
     left: 50%;
     width: ${({ $active }) => ($active ? "80%" : "0")};
     height: 2px;
-    background: linear-gradient(90deg, #d1d1d1ff, #ffffffff);
     border-radius: 2px;
     transform: translateX(-50%);
     transition: width 0.3s ease;
+          background: ${({ $linearGradient }) =>
+        $linearGradient || "linear-gradient(90deg, #d1d1d1ff, #ffffffff)"};
   }
 
   &:hover::after {
@@ -127,13 +100,11 @@ export const LinkElement = styled(Link)<LinkElementProps>`
   }
 
   @media (min-width: ${breakpoints.lg}) {
-    color: ${({ color }) => color || "black"};
     padding: 1.2rem;
 
     &::after {
       width: ${({ $active }) => ($active ? "80%" : "0")};
-      background: ${({ $linearGradient }) =>
-        $linearGradient || "linear-gradient(90deg, #d1d1d1ff, #ffffffff)"};
+
       bottom: 0;
     }
 
@@ -165,7 +136,7 @@ export const MenuContainer = styled.div`
 export const LogoAndMobileMenu = styled.div`
   display: flex;
   justify-content: space-between;
-  width: 100%;
+  width: 90%;
   align-items: center;
 
   @media (min-width: ${breakpoints.lg}) {
